mcp/pyproject.toml
[project]
name = "rnaflow-mcp"
version = "0.2.0"
description = "RNAFlow MCP Server - Supports both local stdio and remote SSE modes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "uvicorn>=0.32.0",
    "starlette>=0.41.0",
    "psutil>=7.2.2",
]

[project.optional-dependencies]
full = [
    "psutil>=6.0.0",
]
